move all egui-related utilities into input/egui
